ci: Remove llvm 5, introduce llvm 16
LLVM 5 is ancient.
