+++ .github/workflows/bindgen.yml
@@ -142,7 +142,7 @@ jobs:
           - debian: null
             cross: null
             rust: null
-        llvm_version: ["5.0", "9.0"]
+        llvm_version: ["9.0", "16.0"]
         main_tests: [1]
         release_build: [0, 1]
         no_default_features: [0, 1]
@@ -155,7 +155,7 @@ jobs:
           # Test with extra asserts + docs just with latest llvm versions to
           # prevent explosion
           - os: ubuntu-latest
-            llvm_version: "9.0"
+            llvm_version: "16.0"
             release_build: 0
             no_default_features: 0
             feature_extra_asserts: 1
@@ -169,17 +169,20 @@ jobs:
           #     debian: arm64
           #     cross: aarch64-linux-gnu
           #     rust: aarch64-unknown-linux-gnu
-          #   llvm_version: "9.0"
+          #   llvm_version: "16.0"
           #   main_tests: 0
           #   release_build: 0
           #   feature_extra_asserts: 0
 
           # Ensure stuff works on macos too
-          - os: macos-latest
-            llvm_version: "9.0"
-            release_build: 0
-            no_default_features: 0
-            feature_extra_asserts: 0
+          # Disabled for now because llvm doesn't provide releases for x86-64
+          # macOS which is what the runner uses.
+          #
+          # - os: macos-latest
+          #   llvm_version: "16.0"
+          #   release_build: 0
+          #   no_default_features: 0
+          #   feature_extra_asserts: 0
     steps:
       - uses: actions/checkout@v3
