+++ CONTRIBUTING.md
@@ -208,8 +208,8 @@ can add multiple test expectations, one for each supported `libclang`
 version. Instead of having a single `bindgen-tests/tests/expectations/tests/my_test.rs` file,
 add each of:
 
+* `bindgen-tests/tests/expectations/tests/libclang-16/my_test.rs`
 * `bindgen-tests/tests/expectations/tests/libclang-9/my_test.rs`
-* `bindgen-tests/tests/expectations/tests/libclang-5/my_test.rs`
 
 If you need to update the test expectations for a test file that generates
 different bindings for different `libclang` versions, you *don't* need to have
@@ -225,12 +225,6 @@ to check the bindings against with a cargo feature:
 ```
 $ cargo test --features __testing_only_libclang_$VERSION
 ```
-
-Where `$VERSION` is one of:
-
-* `4`
-* `3_9`
-* `3_8`
 
 depending on which version of `libclang` you have installed.
