+++ bindgen-cli/Cargo.toml
@@ -40,7 +40,7 @@ which-rustfmt = ["bindgen/which-rustfmt"]
 # Features used for CI testing 
 __testing_only_extra_assertions = ["bindgen/__testing_only_extra_assertions"]
 __testing_only_libclang_9 = ["bindgen/__testing_only_libclang_9"]
-__testing_only_libclang_5 = ["bindgen/__testing_only_libclang_5"]
+__testing_only_libclang_16 = ["bindgen/__testing_only_libclang_16"]
 
 [package.metadata.release]
 release = true
