+++ bindgen-tests/tests/expectations/tests/abi_variadic_function.rs
