+++ bindgen-tests/tests/expectations/tests/auto.rs
