+++ bindgen-tests/tests/expectations/tests/const_bool.rs
