+++ bindgen-tests/tests/expectations/tests/constant-evaluate.rs
