+++ bindgen-tests/tests/expectations/tests/derive-hash-struct-with-incomplete-array.rs
@@ -75,6 +75,8 @@ pub struct test2 {
 }
 #[test]
 fn bindgen_test_layout_test2() {
+    const UNINIT: ::std::mem::MaybeUninit<test2> = ::std::mem::MaybeUninit::uninit();
+    let ptr = UNINIT.as_ptr();
     assert_eq!(
         ::std::mem::size_of::<test2>(),
         4usize,
@@ -84,6 +86,21 @@ fn bindgen_test_layout_test2() {
         ::std::mem::align_of::<test2>(),
         4usize,
         concat!("Alignment of ", stringify!(test2)),
+    );
+    assert_eq!(
+        unsafe { ::std::ptr::addr_of!((*ptr).a) as usize - ptr as usize },
+        0usize,
+        concat!("Offset of field: ", stringify!(test2), "::", stringify!(a)),
+    );
+    assert_eq!(
+        unsafe { ::std::ptr::addr_of!((*ptr).incomplete_array) as usize - ptr as usize },
+        4usize,
+        concat!(
+            "Offset of field: ",
+            stringify!(test2),
+            "::",
+            stringify!(incomplete_array),
+        ),
     );
 }
 #[repr(C)]
@@ -95,6 +112,8 @@ pub struct test3 {
 }
 #[test]
 fn bindgen_test_layout_test3() {
+    const UNINIT: ::std::mem::MaybeUninit<test3> = ::std::mem::MaybeUninit::uninit();
+    let ptr = UNINIT.as_ptr();
     assert_eq!(
         ::std::mem::size_of::<test3>(),
         4usize,
@@ -104,5 +123,32 @@ fn bindgen_test_layout_test3() {
         ::std::mem::align_of::<test3>(),
         4usize,
         concat!("Alignment of ", stringify!(test3)),
+    );
+    assert_eq!(
+        unsafe { ::std::ptr::addr_of!((*ptr).a) as usize - ptr as usize },
+        0usize,
+        concat!("Offset of field: ", stringify!(test3), "::", stringify!(a)),
+    );
+    assert_eq!(
+        unsafe {
+            ::std::ptr::addr_of!((*ptr).zero_length_array) as usize - ptr as usize
+        },
+        4usize,
+        concat!(
+            "Offset of field: ",
+            stringify!(test3),
+            "::",
+            stringify!(zero_length_array),
+        ),
+    );
+    assert_eq!(
+        unsafe { ::std::ptr::addr_of!((*ptr).incomplete_array) as usize - ptr as usize },
+        4usize,
+        concat!(
+            "Offset of field: ",
+            stringify!(test3),
+            "::",
+            stringify!(incomplete_array),
+        ),
     );
 }
