+++ bindgen-tests/tests/expectations/tests/incomplete-array-padding.rs
@@ -122,6 +122,8 @@ pub struct foo {
 }
 #[test]
 fn bindgen_test_layout_foo() {
+    const UNINIT: ::std::mem::MaybeUninit<foo> = ::std::mem::MaybeUninit::uninit();
+    let ptr = UNINIT.as_ptr();
     assert_eq!(
         ::std::mem::size_of::<foo>(),
         8usize,
@@ -131,6 +133,11 @@ fn bindgen_test_layout_foo() {
         ::std::mem::align_of::<foo>(),
         8usize,
         concat!("Alignment of ", stringify!(foo)),
+    );
+    assert_eq!(
+        unsafe { ::std::ptr::addr_of!((*ptr).b) as usize - ptr as usize },
+        8usize,
+        concat!("Offset of field: ", stringify!(foo), "::", stringify!(b)),
     );
 }
 impl Default for foo {
