+++ bindgen-tests/tests/expectations/tests/issue-769-bad-instantiation-test.rs
