--- bindgen-tests/tests/expectations/tests/libclang-5/inline_namespace_nested.rs
@@ -1,1 +1,0 @@
-/* error generating bindings */
