--- bindgen-tests/tests/expectations/tests/libclang-5/issue-2556.rs
@@ -1,1 +1,0 @@
-/* error generating bindings */
