+++ bindgen-tests/tests/expectations/tests/objc_template.rs
