+++ bindgen-tests/tests/expectations/tests/partial-specialization-and-inheritance.rs
