+++ bindgen-tests/tests/expectations/tests/type_alias_template_specialized.rs
