+++ bindgen-tests/tests/expectations/tests/zero-sized-array.rs
@@ -114,6 +114,8 @@ pub struct DynamicallySizedArray {
 }
 #[test]
 fn bindgen_test_layout_DynamicallySizedArray() {
+    const UNINIT: ::std::mem::MaybeUninit<DynamicallySizedArray> = ::std::mem::MaybeUninit::uninit();
+    let ptr = UNINIT.as_ptr();
     assert_eq!(
         ::std::mem::size_of::<DynamicallySizedArray>(),
         0usize,
@@ -124,6 +126,16 @@ fn bindgen_test_layout_DynamicallySizedArray() {
         1usize,
         concat!("Alignment of ", stringify!(DynamicallySizedArray)),
     );
+    assert_eq!(
+        unsafe { ::std::ptr::addr_of!((*ptr).arr) as usize - ptr as usize },
+        0usize,
+        concat!(
+            "Offset of field: ",
+            stringify!(DynamicallySizedArray),
+            "::",
+            stringify!(arr),
+        ),
+    );
 }
 /// No `_address` field here either.
 #[repr(C)]
@@ -133,6 +145,8 @@ pub struct ContainsDynamicallySizedArray {
 }
 #[test]
 fn bindgen_test_layout_ContainsDynamicallySizedArray() {
+    const UNINIT: ::std::mem::MaybeUninit<ContainsDynamicallySizedArray> = ::std::mem::MaybeUninit::uninit();
+    let ptr = UNINIT.as_ptr();
     assert_eq!(
         ::std::mem::size_of::<ContainsDynamicallySizedArray>(),
         0usize,
@@ -142,5 +156,15 @@ fn bindgen_test_layout_ContainsDynamicallySizedArray() {
         ::std::mem::align_of::<ContainsDynamicallySizedArray>(),
         1usize,
         concat!("Alignment of ", stringify!(ContainsDynamicallySizedArray)),
+    );
+    assert_eq!(
+        unsafe { ::std::ptr::addr_of!((*ptr).dsa) as usize - ptr as usize },
+        0usize,
+        concat!(
+            "Offset of field: ",
+            stringify!(ContainsDynamicallySizedArray),
+            "::",
+            stringify!(dsa),
+        ),
     );
 }
